Fixes breaking change to the helpers class method from Sidekiq 8
