+++ lib/sidekiq-status/web/sidekiq-status-single-web.png
