+++ lib/sidekiq-status/web/sidekiq-status-web.png
