+++ lib/sidekiq-status/web/views/status.erb
