+++ lib/sidekiq-status/web/views/status_not_found.erb
