+++ lib/sidekiq-status/web/views/statuses.erb
